replacing the processing recording folder with new one
